myproject/requirements.txt
Django>=5.0.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.5.0
